requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
jinja2>=3.1.0
aiofiles>=23.1.0

opencv-python>=4.8.0
pillow>=9.5.0
pdf2image>=1.16.0
pytesseract>=0.3.10
numpy>=1.21.0

google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.100.0

anthropic>=0.39.0

python-dateutil>=2.8.2
python-dotenv>=1.0.0
